refactor(RelatedProducts): change the entire Implementation of the component and ensure it's rendered dynamically
